requirements.txt
editdistance
bert-serving-server
bert-serving-client
nltk
mosestokenizer
